Reorganized structure of the codebase to make it more easily readible!
