+++ include/renderer/CUDA/CudaBrickPool.cuh
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "renderer/BrickPool.hpp"
+
+using CudaBrickPool = BrickPool<BrickPoolImpl::CudaBrickPoolTraits>;
