+++ src/renderer/CUDA/CudaRenderer.cpp
