+++ src/renderer/CUDA/CudaRenderer_mac_stub.cpp
@@ -1,0 +1,7 @@
+#include "renderer/Renderer.hpp"
+        class CudaRenderer : public Renderer {
+        public:
+            CudaRenderer() {}
+            ~CudaRenderer() {}
+            void Draw(const Character&, unsigned int) override {}
+        };
