+++ src/renderer/D3D12/D3D12Device.cpp
@@ -1,5 +1,5 @@
 #ifdef _WIN32
-#include "renderer/D3D12Device.hpp"
+#include "renderer/D3D12/D3D12Device.hpp"
 #include "State.hpp"
 #include <stdexcept>
 #include <iostream>
